Informes

+++ core/views.py
@@ -1,24 +1,26 @@
 from typing import Any
-from django.shortcuts import redirect
 from django.views.generic import ListView, CreateView, UpdateView
 from django.http import HttpResponse
-from django.urls import reverse
-from .models import Libro
-from .forms import LibroForm
+from .models import Libro, Informe
+from .forms import LibroForm, InformeForm
+from django.contrib.auth.mixins import LoginRequiredMixin
+from usuarios.views import SuperUserRequiredMixin
 
 # Create your views here.
-
 class BusquedaLibros(ListView):
     context_object_name = 'libros'
     paginate_by = 20
 
+    def get_context_data(self, **kwargs: Any) -> dict:
+        return {'titulo': 'Búsqueda de Libros', **super().get_context_data(**kwargs)}
+    
     def get_template_names(self) -> list:
         if(self.request.htmx):
             return 'partials/lista_libros.html'
         
         return 'busqueda_libros.html'
     
-    def filtro_libros(self):
+    def filtro_libros(self, modelo = Libro):
         autor = self.request.GET.get('autor', None)
         titulo = self.request.GET.get('titulo', None)
         descriptores = self.request.GET.get('descriptores', None)
@@ -27,11 +29,11 @@ class BusquedaLibros(ListView):
         libros, prev_libros = None, None
 
         if(autor):
-            libros = Libro.objects.filter(autores__icontains=autor)
+            libros = modelo.objects.filter(autores__icontains=autor)
             prev_libros = libros
 
         if(titulo):
-            libros = Libro.objects.filter(titulo__icontains=titulo) if not libros else libros.filter(titulo__icontains=titulo)
+            libros = modelo.objects.filter(titulo__icontains=titulo) if not libros else libros.filter(titulo__icontains=titulo)
         
         if(libros):
             prev_libros = libros
@@ -39,7 +41,7 @@ class BusquedaLibros(ListView):
             libros = prev_libros
 
         if(ano):
-            libros = Libro.objects.filter(ano_publicacion__icontains=ano) if not libros else libros.filter(ano_publicacion__icontains=ano)
+            libros = modelo.objects.filter(ano_publicacion__icontains=ano) if not libros else libros.filter(ano_publicacion__icontains=ano)
 
         if(libros):
             prev_libros = libros
@@ -49,7 +51,7 @@ class BusquedaLibros(ListView):
         if(descriptores):
             for descriptor in descriptores.split(';'):
                 descriptor = descriptor.strip()
-                libros = Libro.objects.filter(descriptores__nombre__icontains=descriptor) if not libros else libros.filter(descriptores__nombre__icontains=descriptor)
+                libros = modelo.objects.filter(descriptores__nombre__icontains=descriptor) if not libros else libros.filter(descriptores__nombre__icontains=descriptor)
 
         if(not libros):
             libros = prev_libros
@@ -67,7 +69,7 @@ class BusquedaLibros(ListView):
             if(not libros and len(self.request.GET.keys())):
                 libros = Libro.objects.none()
         
-        if((not self.request.htmx or all(y == '' for y in self.request.GET.values() )) and not libros):
+        if((not self.request.htmx or all(y == '' for x,y in self.request.GET.items() if x != 'page' )) and not libros):
             libros = Libro.objects.all()
 
         if(libros):
@@ -75,12 +77,15 @@ class BusquedaLibros(ListView):
 
         return libros
 
-class CreacionLibro(CreateView):
+class CreacionLibro(SuperUserRequiredMixin, CreateView):
     template_name = 'libro_form.html'
     form_class = LibroForm
     success_url = '/'
 
-class EdicionLibro(UpdateView):
+    def get_context_data(self, **kwargs: Any) -> dict:
+        return {'titulo': 'Registro de Nuevo Libro', **super().get_context_data(**kwargs)}
+
+class EdicionLibro(SuperUserRequiredMixin, UpdateView):
     template_name = 'libro_form.html'
     form_class = LibroForm
     success_url = '/'
@@ -94,8 +99,90 @@ class EdicionLibro(UpdateView):
         return context
     
 def eliminar_libro(request, pk):
-    if(request.method == 'POST'):
+    if(request.method == 'POST' and request.user.is_superuser):
         libro = Libro.objects.get(id=pk)
         libro.delete()
 
-    return HttpResponse("")
+        return HttpResponse(status=200)
+    
+    return HttpResponse(status=402)
+
+class BusquedaInformes(LoginRequiredMixin, BusquedaLibros):
+    context_object_name = 'informes'
+    paginate_by = 20
+
+    def get_template_names(self) -> list:
+        if(self.request.htmx):
+            return 'partials/lista_informes.html'
+        
+        return 'busqueda_informes.html'
+    
+    def get_context_data(self, **kwargs: Any) -> dict:
+        return {'titulo': 'Búsqueda de Informes Técnicos', **super().get_context_data(**kwargs)}
+    
+    def filtro_libros(self, modelo = Informe):
+        informes = super().filtro_libros(modelo)
+
+        solicitud_servicio = self.request.GET.get('solicitud_servicio', None)
+        programa = self.request.GET.get('programa', None)
+
+        prev_informes = informes
+        if(solicitud_servicio):
+            informes = modelo.objects.filter(solicitud_servicio__icontains=solicitud_servicio) if not informes else informes.filter(solicitud_servicio__icontains=solicitud_servicio)
+
+        if(not informes):
+            informes = prev_informes
+
+        if(programa):
+            informes = modelo.objects.filter(programa__nombre__icontains=programa) if not informes else informes.filter(programa__nombre__icontains=programa)
+
+        if(not informes):
+            informes = prev_informes
+
+        return informes
+    
+    def get_queryset(self):
+        informes = None
+
+        if(self.request.htmx and len(self.request.GET.keys())):
+            informes = self.filtro_libros()
+
+            if(not informes and len(self.request.GET.keys())):
+                informes = Informe.objects.none()
+        
+        if((not self.request.htmx or all(y == '' for x,y in self.request.GET.items() if x != 'page')) and not informes):
+            informes = Informe.objects.all()
+
+        if(informes):
+            informes.prefetch_related('descriptores')
+
+        return informes
+
+class CreacionInforme(SuperUserRequiredMixin, CreateView):
+    template_name = 'informe_form.html'
+    form_class = InformeForm
+    success_url = 'publicaciones/busqueda/informes/'
+
+    def get_context_data(self, **kwargs: Any) -> dict:
+        return {'titulo': 'Registro de Informe Técnico', **super().get_context_data(**kwargs)}
+
+class EdicionInforme(SuperUserRequiredMixin, UpdateView):
+    template_name = 'informe_form.html'
+    form_class = InformeForm
+    success_url = '/publicaciones/busqueda/informes/'
+    model = Informe
+
+    def get_context_data(self, **kwargs: Any) -> dict:
+        context = super().get_context_data(**kwargs)
+        context['titulo'] = 'Edición de Informe Técnico'
+        context['form'].fields['descriptores'].initial = "; ".join(context['object'].descriptores.all().order_by('nombre').values_list('nombre', flat=True))
+
+        return context
+    
+def eliminar_informe(request, pk):
+    if(request.method == 'POST' and request.user.is_superuser):
+        informe = Informe.objects.get(id=pk)
+        informe.delete()
+        return HttpResponse(status=200)
+    
+    return HttpResponse(status=402)
